Refactor into modular package with CI testing infrastructure
Convert from 4 flat top-level modules to a proper tomoe/ package:
- Deduplicate check_port_open (3x), auth username construction (8x),
  WinRM client kwargs (7x), and auth error pattern matching (5x)
- Unify error hierarchy: TomoeError > AuthenticationError, ConnectionError
- Collapse 12-branch protocol×operation dispatch into 3-branch dispatch
  via protocol module dict
- Add full test suite: 36 unit tests + 16 integration tests
- Add GitHub Actions CI: unit tests, SSH integration (Docker),
  WinRM+SMB integration (Windows runner)
