+++ tomoe/__main__.py
@@ -1,0 +1,3 @@
+from tomoe.cli import main
+
+main()
